Use the conda-build 3 compiler syntax
Switches from using the `toolchain` to the `conda-build` 3 `compiler`
syntax. This still uses `toolchain` under the hood while allowing us to
move to a syntax that is more inline with `AnacondaRecipes`. Also
provides the compiler run time as a dependency on Linux and Windows
(macOS still pending on this change).