Store extra C flags in new environment variable
Instead of just overwriting the `CFLAGS`, store the additional `CFLAGS`
in a new environment variable and append them. It's likely that `CFLAGS`
was set this way before `toolchain` was added and so assuming `CFLAGS`
to be empty was reasonable and common place at that time. However that
is no longer appropriate with the `toolchain` or the new compilers. So
this corrects this behavior to match modern building conventions.